requirements.txt
streamlit
torch
transformers
langchain
pillow
accelerate
scikit-learn
